requirements.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
greenlet==3.0.1
idna==3.4
numpy==1.26.2
pandas==2.1.3
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyWeatherKit==0.0.4
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
